restructure utils file
